support AllNamespaces operators; some script cleanup
There was a bug in the script that it was not using the
TARGET_NAMESPACE.
Some operators watch for CRs in all namespaces - they use
installModes AllNamespaces - when creating the operatorgroup,
we have to use an empty spec `{}` so that it will apply to all
namespaces.

Run the script with `set -euo pipefail` and cleanup some of the
issues found.